requirements.txt
# Core dependencies for dog emotion recognition ML package
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
scipy>=1.7.0
joblib>=1.1.0

# Ensemble learning algorithms
xgboost>=1.5.0
lightgbm>=3.3.0

# Data processing and visualization
openpyxl>=3.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Computer vision and image processing
opencv-python>=4.5.0
Pillow>=8.0.0

# Deep learning frameworks for classification algorithms
torch>=1.12.0
torchvision>=0.13.0
timm>=0.6.0
transformers>=4.20.0

# YOLO detection (optional for tail detection)
# ultralytics>=8.0.0

# YAML processing for Roboflow data.yaml
PyYAML>=6.0

# Optional dependencies for advanced features
# Install with: pip install -r requirements.txt 
cursor-notebook-mcp==0.2.3